pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.2"}
uvicorn = "^0.30.6"
biopython = "^1.84"
pydantic = "^2.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
